requirements.txt
Django==2.0.3

djangorestframework==3.8.2
djangorestframework_jwt==1.11.0
django-cors-headers==2.2.0
django-rest-swagger==2.2.0
django-storages==1.6.6
google-cloud-storage>=0.22.0

django-grappelli==2.11.1
django-ckeditor==5.4.0
django2.0-super-inlines==0.1.5

django-debug-toolbar==1.9


psycopg2-binary==2.7.4

wheel==0.30.0
gunicorn==19.7.1

coverage==4.5.1

pyyaml==3.12
